ContextRecord refactoring
